Modified the structure of  each page in AllTier to accommodate amplitude tracking.
page tsx files are for metadata, client pages are for the page content

Added amplitude tracking to login and sign up page.

Changed DecodedToken 0 decoded from JWT access token
