+++ src/app/layout.tsx
@@ -6,8 +6,11 @@ import { AuthProvider } from "@/components/AuthProvider";
 
 // Now you can safely export metadata because this is a Server Component.
 export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+  title: {
+    default: "AllTier",   // fallback
+    template: "%s | AllTier",  // lets pages inject titles dynamically
+  },
+  description: "%s | AllTier Page Description"
 };
 
 export default function RootLayout({
